Generates an R script with the examples from an Rd file.

Generates an R script with the examples from an Rd file.

generate_script_from_help(path_to_rd)

Arguments

path_to_rd

String. Path to Rd file you wish to convert to a script

Value

Side-effect. No returned object, writes a script on disk.

Details

This function is a wrapper around `tools::Rd2ex()` tht extracts the examples from Rd files included in the man/ folder of source packages and writes them to disk, next to the Rd files.

Examples

get_archived_sources("AER")
#> # A tibble: 25 × 4
#>    version   url                                       last_modified       size 
#>    <chr>     <chr>                                     <dttm>              <chr>
#>  1 AER_0.2-1 https://cran.r-project.org/src/contrib/A… 2008-05-05 00:22:00 2.4M 
#>  2 AER_0.2-2 https://cran.r-project.org/src/contrib/A… 2008-05-05 11:43:00 2.4M 
#>  3 AER_0.9-0 https://cran.r-project.org/src/contrib/A… 2008-05-28 23:21:00 2.4M 
#>  4 AER_1.0-0 https://cran.r-project.org/src/contrib/A… 2008-08-26 12:53:00 2.6M 
#>  5 AER_1.0-1 https://cran.r-project.org/src/contrib/A… 2008-10-22 13:15:00 2.6M 
#>  6 AER_1.1-0 https://cran.r-project.org/src/contrib/A… 2009-02-05 19:44:00 2.6M 
#>  7 AER_1.1-1 https://cran.r-project.org/src/contrib/A… 2009-03-08 15:32:00 2.6M 
#>  8 AER_1.1-2 https://cran.r-project.org/src/contrib/A… 2009-03-19 15:59:00 2.6M 
#>  9 AER_1.1-3 https://cran.r-project.org/src/contrib/A… 2009-05-22 00:31:00 2.6M 
#> 10 AER_1.1-4 https://cran.r-project.org/src/contrib/A… 2009-09-23 22:49:00 2.7M 
#> # … with 15 more rows