Downloads Rd from an archived source package

Downloads Rd from an archived source package

get_man_package(name, version, url, clean = TRUE)

Arguments

name

String. Name of the packages.

version

String. Version of the package to download.

url

String. Link to archived package tar file.

clean

Boolean, defaults to TRUE. If TRUE, only keeps man/ folder containing the documentaton. If FALSE, keeps entire package.

Value

Side-effect. No returned object, writes a Rd files to disk.

Details

This function returns a data frame with a column `name` giving the name of a package, `version` giving its version, `url` the download url `last_modified` the date on which the package was last modified on CRAN and `size`, the size of the package

Examples

if (FALSE) {
# First, get list of packages. In this case, the ones in the "Econometrics" view as of "2015-01-01"
ctv_econ <- get_packages_from_view("Econometrics", date = "2015-01-01")

# Now, get the names, versions, and urls for these packages
ctv_econ_sources <- get_sources_for_selected_packages(ctv_econ)

# It is now possible to download the man/ folders of these packages with the following lines
ctv_econ_sources %>%
  mutate(get_sources = pmap(list(name, version, url), get_man_package))
}